feat(sharing): edit shared boards in place + target them from charts
Makes shared dashboards editable (previously create/rename/visibility only):

- Extract EditableDashboard: the section/widget grid + add/remove/move/section
  controls, shared by the home Dashboard and shared boards (Dashboard refactored
  to use it; behaviour unchanged).
- /u/:username/:slug: when the logged-in owner views their own board (token +
  username === auth.user), render the live, editable dashboard with an Edit
  toggle, saving via updateSharedDashboard. Non-owners keep the read-only
  public view.
- Chart "Add to dashboard": add a target-dashboard selector (home or any shared
  board) before the section selector; saves to the chosen board.
- docs: document editing + chart targeting.
