+++ docs/features/sharing.md
@@ -63,8 +63,15 @@ beyond the saved widgets.
 - **View** a shared dashboard at `/u/<username>/<slug>` and a profile at
   `/u/<username>`. These pages are public, render without the app chrome
   (header/sidebar/footer), and fetch nothing per widget — they render from the
-  server-resolved data. Per-widget editing of a shared dashboard (beyond seeding
-  from the home dashboard) is a planned follow-up.
+  server-resolved data.
+- **Edit in place**: when you are logged in and viewing your *own* shared
+  dashboard (`/u/<you>/<slug>`), an Edit toggle appears and you get the same
+  add/remove/move-widget and section controls as the home dashboard; changes
+  save to that shared dashboard. (Owners see live widget data here, not the
+  read-only snapshot.)
+- **Add a chart to a shared dashboard**: the chart page's "Add to dashboard"
+  dialog lets you pick the target dashboard (your home dashboard or any shared
+  one) and then the section.
 
 ## API
